requirements.txt
# Inventory Management System - Dependencies
# ==========================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Excel/CSV Processing
openpyxl==3.1.2
pandas==2.1.4

# HTTP Client (for inter-service calls)
httpx==0.26.0

# Redis (for caching & message queue)
redis==5.0.1

# Background Tasks
celery==5.3.6

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.0
isort==5.13.2




